+++ .github/workflows/nightly-build-artifact.yml
@@ -116,7 +116,7 @@ jobs:
         with:
           version: '22.3.1'
           java-version: '17'
-          components: 'native-image'
+          components: 'espresso,native-image'
           github-token: ${{ secrets.GITHUB_TOKEN }}
           cache: 'maven'
       - uses: docker/login-action@v2
