+++ distribution/proxy-native/src/main/release-docs/LICENSE
@@ -357,4 +357,13 @@ The text of each license is also included at licenses/LICENSE-[project].txt.
     checker-qual 3.5.0: https://github.com/typetools/checker-framework/blob/master/checker-qual, MIT
     jul-to-slf4j 1.7.36: https://www.slf4j.org, MIT
     slf4j-api 1.7.36: https://www.slf4j.org, MIT
-    jnanoid 2.0.0: https://github.com/aventrix/jnanoid, MIT 
+    jnanoid 2.0.0: https://github.com/aventrix/jnanoid, MIT
+
+========================================================================
+UPL licenses
+========================================================================
+
+The following components are provided under the UPL License. See project link for details.
+The text of each license is also included at licenses/LICENSE-[project].txt.
+
+    truffle-api 22.3.1: http://www.graalvm.org/, UPL
