+++ distribution/proxy/src/main/release-docs/LICENSE
@@ -360,3 +360,12 @@ The text of each license is also included at licenses/LICENSE-[project].txt.
     jul-to-slf4j 1.7.36: https://www.slf4j.org, MIT
     slf4j-api 1.7.36: https://www.slf4j.org, MIT
     jnanoid 2.0.0: https://github.com/aventrix/jnanoid, MIT 
+
+========================================================================
+UPL licenses
+========================================================================
+
+The following components are provided under the UPL License. See project link for details.
+The text of each license is also included at licenses/LICENSE-[project].txt.
+
+    truffle-api 21.2.0: http://www.graalvm.org/, UPL
