+++ infra/util-groovy/pom.xml
@@ -21,23 +21,20 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.apache.shardingsphere</groupId>
-        <artifactId>shardingsphere</artifactId>
+        <artifactId>shardingsphere-infra</artifactId>
         <version>5.3.2-SNAPSHOT</version>
     </parent>
-    <artifactId>shardingsphere-infra</artifactId>
-    <packaging>pom</packaging>
-    <name>${project.artifactId}</name>
     
-    <modules>
-        <module>util</module>
-        <module>common</module>
-        <module>parser</module>
-        <module>binder</module>
-        <module>route</module>
-        <module>rewrite</module>
-        <module>executor</module>
-        <module>merge</module>
-        <module>context</module>
-        <module>util-groovy</module>
-    </modules>
+    <artifactId>shardingsphere-infra-util-groovy</artifactId>
+    
+    <dependencies>
+        <dependency>
+            <groupId>org.apache.groovy</groupId>
+            <artifactId>groovy</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.google.guava</groupId>
+            <artifactId>guava</artifactId>
+        </dependency>
+    </dependencies>
 </project>
